call CorrectGeometry when outputting
This fixes the funky landcover/landuse issue, but increases runtime
to 2m50s.

An option to mitigate the extra 11s: if the geometry had to be
corrected, store it in the mmap store vs re-creating on the fly.

The larger slowdown of ~17s (2m22->2m39s) is probably from creating
objects on the fly? Perhaps we could move to handing out shared_ptrs
and keep a z6-specific cache.

Let's get relations into the system first, then tackle that.