extract existing RelationStore to own files
It looks like this maybe wasn't finished - nothing seems
to actually insert things into it.

Next step is to get it to the level of the other BinarySearchXyzStores,
and re-write it to store only way IDs, not full node lists.

It won't be as memory efficient as a SortedRelationStore, but I think
it'll be enough to get going on. There's generally a factor of 10 fewer
relations than ways (and 10 fewer ways than nodes), so the benefit of
a SortedRelationStore is hopefully not as urgently needed.