support geometries from relations
Memory usage for GB isn't improved -- I think this is because
ReadPhase::Relations ends up with the biggest relations all in memory
simultaneously anyway.

If I run with fewer threads, memory usage is improved--so I think
this ought to reduec memory usage for much larger PBFs, like NA
or Germany.