try storing relation geoms in geom store
...and only doing lazy geom for way geoms

rationale: relations aren't as much of a working set
memory savings as ways, since ways are often limited to a
single z14 tile

NS: 22s

GB: 2m39s

Note that this isn't a very fair comparison, as I've pessimized
buildWayGeometry - it always does a copy to ensure we're operating
on a geometry that uses the std allocator.

I think to give it a fair test, we'd need to make a zero cost
abstraction that has no penalty for mmap allocated objects. I think
that's doable, but the best way I can think of at the moment involves
duplicating some code. Let's shelve this until I've finished with
other code changes in this area.