extract ClipCache to own file, be pickier on use
Some housekeeping: extract clip_cache.cpp

And a small tweak: only use clip cache for large polygons.

Previously, we ran all polygons through the clip cache. But the clip
cache is intended for polygons that span multiple tiles.

For the most part, the old code worked well enough -- small polygons only
occupied a single z14 tile, and so would fail the test that checked
if we were adding at the terminal zoom.

Still, we would always check the clip cache, even for very small
polygons. Even though a map lookup is O(1), it's still slow when
done at scale.

Instead, when we add the geometry to the index, we make a note of
whether it's large or not, and encode that in its object ID.

I'm on the fence about whether this tweak is good. It adds complexity
for an inconsistent speedup. I think it'll be more useful when
scaling to more cores, as it should reduce contention on the
clipCache locks.

master is 2m22s for GB, this can sometimes hit 2m20s, but sometimes
also 2m22s, suggesting it's not a slam dunk.