reuse vector for finalizeOutputs
Another one that I'm not at all confident in. The idea is that the
vector will grow to its terminal max size, so we'll stop paying
costs to alloc memory.

With this, I seem to get 2m20s/2m21s for GB.

Starting to think: extracting clip_cache to its own file is good,
everything else is bad.