experiments in simplifying
We currently simplify from a clip of the raw polygon. My intuition is
that this must be expensive -- a z6 simplification starts off with a
polygon suitable for z14, which is far more data than is needed.

This commit adds some tools to test ideas like:

- is simplify(simplify(geom, X), X) faster than simplify(geom, X*2) ?
- is simplify(simplify(geom, X), X) similar to simplify(geom, X*2) ?

I think my conclusion is:

- yes, it's faster to simplify a simplified geometry
- it's...similar. I don't know enough of the history of simplification
  to know if it's similar enough.

It'd be fastest to maintain all the simplifications, eg:

- simplify(geo, X)
- simplify(simplify(geo, X), X*2)
- simplify(simplify(simplify(geo, X), X*2), X*4)

But in practice, even the first simplification alone gives you a huge boost for
complex multipolygons.

Perhaps it'd be reasonable to use the usual mechanism, except for
zooms below some zoom, we cache a simplified polygon that we use in
place of the raw polygon? It's all a little unclear.

Testing with the GSMNP relation
(https://www.openstreetmap.org/relation/2131838):

- simplifying the raw polygon to 3 different levels: 49ms
- simplifying a simplified polygon to 3 different levels: 18ms