wip: an experiment in simplifying
The idea - could we save time without impacting the map if we
simplified a geometry before caching it in the clip cache.

Simplification takes ~13% of runtime, e.g. if I set `simplify_below`
to `0`, it takes 1m45s to build GB. (...but the tiles will be much
larger, and rendering is slower.)

this branch, with --lazy-geometries (needed to identify relation vs
way):
real	1m58.290s
user	28m0.471s
sys	0m15.035s

lua-interop with --lazy-geometries:
real	2m0.163s
user	28m46.487s
sys	0m13.853s

Only a 1.5% decrease in runtime -- definitely not worth it...yet.

Possible future idea: when storing a relation, simplify it then and
there and store the simplified version. Due to LeasedStore being
per-thread, we'd be guaranteed that it's located at ID N+1.

pro: we'd never have low-zooms racing to simplify the same geometry

con: more virtual memory needed, we'd compute a simplification for
every relation, even if it never was used in a simplified form